Replaced fabric8 go executable with a bash script.
This commit replaces the go executable with a bash script. The
script makes the image much lighter (50Mb -> 18Mb) as well as
make it a bit more cross platform -- runs anywhere where bash
runs and you don't need to rely on specifying the correct
executable for your environment.

The bash script also includes a few additional features, such
as waiting for a TCP service and specifying the connect timeout.

This image can therefore be used as a drop in replacement for
https://github.com/fabric8-services/fabric8-dependency-wait-service
directly.